BFT-496: Move the runner under the executor
