+++ node/actors/network/src/gossip/mod.rs
@@ -12,10 +12,7 @@
 //! Static connections constitute a rigid "backbone" of the gossip network, which is insensitive to
 //! eclipse attack. Dynamic connections are supposed to improve the properties of the gossip
 //! network graph (minimize its diameter, increase connectedness).
-pub use self::attestation_status::{
-    AttestationStatusClient, AttestationStatusReceiver, AttestationStatusRunner,
-    AttestationStatusWatch, LocalAttestationStatusClient,
-};
+pub use self::attestation_status::{AttestationStatusReceiver, AttestationStatusWatch};
 pub use self::batch_votes::BatchVotesPublisher;
 use self::batch_votes::BatchVotesWatch;
 use crate::{gossip::ValidatorAddrsWatch, io, pool::PoolWatch, Config, MeteredStreamStats};
